Move the phenopacket IO back to `lirical-io`.
