+++ lirical-cli/pom.xml
@@ -46,30 +46,6 @@
             <artifactId>svart</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.phenopackets</groupId>
-            <artifactId>phenopacket-schema</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.phenopackets.phenopackettools</groupId>
-            <artifactId>phenopacket-tools-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.phenopackets.phenopackettools</groupId>
-            <artifactId>phenopacket-tools-io</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.phenopackets.phenopackettools</groupId>
-            <artifactId>phenopacket-tools-util</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.protobuf</groupId>
-            <artifactId>protobuf-java</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.google.protobuf</groupId>
-            <artifactId>protobuf-java-util</artifactId>
-        </dependency>
-        <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
         </dependency>
