+++ lirical-io/src/main/java/org/monarchinitiative/lirical/io/analysis/PhenopacketImporter.java
@@ -1,7 +1,10 @@
-package org.monarchinitiative.lirical.cli.pp;
+package org.monarchinitiative.lirical.io.analysis;
 
 import java.io.InputStream;
 
+/**
+ * Decode phenopacket data from an input stream.
+ */
 public interface PhenopacketImporter {
 
     PhenopacketData read(InputStream is) throws PhenopacketImportException;
