+++ lirical-io/src/main/java/org/monarchinitiative/lirical/io/analysis/PhenopacketImporters.java
@@ -1,5 +1,8 @@
-package org.monarchinitiative.lirical.cli.pp;
+package org.monarchinitiative.lirical.io.analysis;
 
+/**
+ * Get {@link PhenopacketImporter}s to ingest v1 or v2 phenopackets.
+ */
 public class PhenopacketImporters {
 
     private PhenopacketImporters() {
