+++ lirical-io/src/main/java/org/monarchinitiative/lirical/io/analysis/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Read user input formatted as a v1 or v2 phenopacket.
  */
-package org.monarchinitiative.lirical.cli.pp;
+package org.monarchinitiative.lirical.io.analysis;
