+++ lirical-io/src/test/resources/org/monarchinitiative/lirical/io/analysis/BBS1.json
