+++ lirical-io/src/test/resources/org/monarchinitiative/lirical/io/analysis/pfeiffer.v2.yaml
