+++ lirical-io/src/test/resources/org/monarchinitiative/lirical/io/analysis/pfeifferNoVcf.json
